# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Citas Médicas",
        "description": "Documentación de la API",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/consultorios": {
            "get": {
                "tags": [
                    "Consultorios"
                ],
                "summary": "Lista todos los consultorios",
                "operationId": "4128008a2bda4165021f87ee6a43eff5",
                "responses": {
                    "200": {
                        "description": "Lista de consultorios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Consultorio"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Consultorios"
                ],
                "summary": "Crear un nuevo consultorio",
                "operationId": "ef5e542fbe3dfbdf6f26bbb18806b1e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consultorio"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consultorio creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consultorio"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/consultorios/{id}": {
            "get": {
                "tags": [
                    "Consultorios"
                ],
                "summary": "Muestra un consultorio específico",
                "operationId": "10deb26a7d79d6797165b287f4d3e6f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consultorio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del consultorio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consultorio"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consultorio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Consultorios"
                ],
                "summary": "Actualizar un consultorio existente",
                "operationId": "0bb02eb122cc136699f48610363392c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consultorio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consultorio"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consultorio actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consultorio"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "404": {
                        "description": "Consultorio no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Consultorios"
                ],
                "summary": "Eliminar un consultorio",
                "operationId": "ff9339ecee74f85025c4f1b3f187a9a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consultorio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consultorio eliminado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consultorio no encontrado"
                    }
                }
            }
        },
...
        "/api/AgregarDoctor": {
            "post": {
                "tags": [
                    "Doctores"
                ],
                "summary": "Crea un nuevo doctor",
                "operationId": "fa354e6be148ef04efd8b510f40c155b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Doctor"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doctor creado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/doctor/{id}": {
            "get": {
                "tags": [
                    "Doctores"
                ],
                "summary": "Obtiene un doctor por ID",
                "operationId": "7ae317a48767d61dfc829cb5fcd0a285",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor no encontrado"
                    }
                }
            }
        },
        "/api/ActualizarDoctor/{id}": {
            "put": {
                "tags": [
                    "Doctores"
                ],
                "summary": "Actualiza un doctor por ID",
                "operationId": "3b5a0c4bb6d95806f3ab22e1bf540224",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Doctor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctor actualizado"
                    },
                    "404": {
                        "description": "Doctor no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/EliminarDoctor/{id}": {
            "delete": {
                "tags": [
                    "Doctores"
                ],
                "summary": "Elimina un doctor por ID",
                "operationId": "bca72af6f277952c3d9babacb1574b68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor eliminado"
                    },
                    "404": {
                        "description": "Doctor no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Consultorio": {
                "title": "Consultorio",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "ubicacion": {
                        "type": "string",
                        "nullable": true
                    },
                    "piso": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Doctor": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "apellido": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "telefono": {
                        "type": "string",
                        "nullable": true
                    },
                    "genero": {
                        "type": "string",
                        "nullable": true
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "activo",
                            "inactivo"
                        ]
                    },
                    "especialidad_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}